+++ src/Statement/ValuesStatement.php
@@ -3,11 +3,12 @@
 /**
  * @author Jared King <j@jaredtking.com>
  *
- * @link http://jaredtking.com
+ * @see http://jaredtking.com
  *
  * @copyright 2015 Jared King
  * @license MIT
  */
+
 namespace JAQB\Statement;
 
 class ValuesStatement extends Statement
@@ -15,7 +16,7 @@ class ValuesStatement extends Statement
     /**
      * @var array
      */
-    protected $insertValues = [];
+    protected $insertRows = [];
 
     /**
      * Adds values to the statement.
@@ -24,19 +25,47 @@ class ValuesStatement extends Statement
      */
     public function addValues(array $values)
     {
-        $this->insertValues = array_replace($this->insertValues, $values);
+        // Check if this is a multi-dimensional array
+        $isMultiDimensional = 0 == count($values) || (isset($values[0]) && is_array($values[0]));
+
+        if ($isMultiDimensional) {
+            $this->insertRows = array_merge($this->insertRows, $values);
+        } else {
+            // If a single row is provided then it is added
+            // to the last row, or a new row if there are none.
+            // This is done to maintain BC
+            if (count($this->insertRows) > 0) {
+                $k = count($this->insertRows) - 1;
+                $this->insertRows[$k] = array_replace($this->insertRows[$k], $values);
+            } else {
+                $this->insertRows = [$values];
+            }
+        }
 
         return $this;
     }
 
     /**
-     * Gets the values being inserted.
+     * @deprecated
+     * Gets the values being inserted (first row only).
+     *
+     * WARNING: this only returns the first row for BC
      *
      * @return array
      */
     public function getInsertValues()
     {
-        return $this->insertValues;
+        return count($this->insertRows) > 0 ? $this->insertRows[0] : [];
+    }
+
+    /**
+     * Gets the rows being inserted.
+     *
+     * @return array
+     */
+    public function getInsertRows()
+    {
+        return $this->insertRows;
     }
 
     public function build()
@@ -44,20 +73,32 @@ class ValuesStatement extends Statement
         // reset the parameterized values
         $this->values = [];
 
+        // get the list of fields from the first row
+        $keys = [];
         $fields = [];
-        foreach ($this->insertValues as $key => $value) {
-            if ($id = $this->escapeIdentifier($key)) {
-                $fields[] = $id;
-                $this->values[] = $value;
+        if (count($this->insertRows) > 0) {
+            foreach (array_keys($this->insertRows[0]) as $key) {
+                if ($id = $this->escapeIdentifier($key)) {
+                    $fields[] = $id;
+                    $keys[] = $key;
+                }
             }
         }
 
-        if (count($fields) == 0) {
+        if (0 == count($fields)) {
             return '';
         }
 
-        // produces "(`col1`,`col2`,`col3`) VALUES (?,?,?)"
-        return '('.implode(', ', $fields).') VALUES ('.
-            implode(', ', array_fill(0, count($fields), '?')).')';
+        foreach ($this->insertRows as $row) {
+            foreach ($keys as $key) {
+                $this->values[] = array_value($row, $key);
+            }
+        }
+
+        $rowPlaceholders = '('.implode(', ', array_fill(0, count($fields), '?')).')';
+
+        // produces "(`col1`,`col2`,`col3`) VALUES (?,?,?), (?,?,?)"
+        return '('.implode(', ', $fields).') VALUES '
+            .implode(', ', array_fill(0, count($this->insertRows), $rowPlaceholders));
     }
 }